requirements.txt
Django==3.1.5
gunicorn==20.0.4
psycopg2==2.8.6
django-heroku==0.3.1
djangorestframework-simplejwt==4.6.0
pytest-django==4.1.0
